Update script to ignore dll/exe files that do not have accompanying pdb files
D3DFMIQ-2582


Former-commit-id: e0fa33a5381fa4d2ee47d494833e7cd163084b60

+++ tools/gather_for_infer.py
@@ -34,8 +34,10 @@ def copy_dll(destination_directory: Path, src_dll_path: Path, repo_path: Path) -
     goal_path = destination_directory / relative_path
     goal_path.parent.mkdir(parents=True, exist_ok=True)
 
-    shutil.move(str(src_dll_path), str(goal_path))
-    shutil.move(str(src_dll_path.with_suffix('.pdb')), str(goal_path.with_suffix('.pdb')))
+    src_pdb_path = src_dll_path.with_suffix('.pdb')
+    if src_pdb_path.exists() and src_dll_path.exists() :
+        shutil.move(str(src_dll_path), str(goal_path))
+        shutil.move(str(src_pdb_path), str(goal_path.with_suffix('.pdb')))
 
 
 def run(src_directories: List[Path], bin_directory: Path,destination_directory: Path) -> None:
